+++ app/Http/Controllers/DatController.php
@@ -18,13 +18,6 @@ class DatController extends Controller
     private const PATH_IN = 'data/in';
 
     /**
-     * The storage path to `.done.dat` files.
-     *
-     * @var string
-     */
-    private const PATH_OUT = 'data/out';
-
-    /**
      * Create a new controller instance.
      *
      * @return void
@@ -47,11 +40,7 @@ class DatController extends Controller
      */
     public function list()
     {
-        return response()->json(
-            Dat::with('user')
-            ->get()
-            ->map([Dat::class, 'toArray'])
-        );
+        return response()->json(Dat::with('user')->get());
     }
 
     /**
@@ -63,11 +52,13 @@ class DatController extends Controller
     public function read($id)
     {
         /**
-         * @var Dat $dat
+         * The Dat file.
+         *
+         * @var Dat|null $dat
          */
         $dat = Dat::find($id);
 
-        if(!$dat?->file)
+        if(!$dat || $dat->deleted_at)
         {
             return response()->json(['error' => 'File not found'], 404);
         }
@@ -147,6 +138,8 @@ class DatController extends Controller
             {
                 $readFile['error'] = 'Could not save file';
 
+                $code = 500;
+
                 $readFiles[] = $readFile;
 
                 continue;
@@ -156,9 +149,11 @@ class DatController extends Controller
             {
                 $readFile['error'] = 'File was not stored';
 
+                $code = 500;
+
                 $readFiles[] = $readFile;
 
-                $dat->delete();
+                $dat->forceDelete();
 
                 continue;
             }
@@ -171,5 +166,101 @@ class DatController extends Controller
         }
 
         return response()->json($readFiles, $code);
+    }
+
+    /**
+     * Update the current `.dat` file referenced by the `$id`.
+     *
+     * @param int $id
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function update($id)
+    {
+        /**
+         * The Dat file.
+         *
+         * @var Dat|null $dat
+         */
+        $dat = Dat::find($id);
+
+        if(!$dat || $dat->deleted_at)
+        {
+            return response()->json(['error' => 'File not found'], 404);
+        }
+
+        /**
+         * The files on the Request.
+         *
+         * Only the first one is needed.
+         *
+         * @var \Symfony\Component\HttpFoundation\FileBag $files
+         */
+        $files = request()->files;
+
+        /**
+         * The new `.dat` file.
+         *
+         * @var UploadedFile $file
+         */
+        $file = $files->get(array_key_first($files->all()));
+
+        if(!Dat::validate($file))
+        {
+            return response()->json(['error' => 'Invalid file'], 422);
+        }
+
+        /**
+         * The local storage instance.
+         *
+         * @var \Illuminate\Support\Facades\Storage $storage
+         */
+        $storage = app('storage')::disk('local');
+
+        if(!$storage->putFileAs(self::PATH_IN, $file, $dat->name))
+        {
+            return response()->json(['error' => 'File was not stored'], 500);
+        }
+
+        $dat->updated_at = Carbon::now();
+
+        if(!$dat->save())
+        {
+            return response()->json(['error' => 'File was not saved'], 500);
+        }
+
+        return response()->json($dat->toArray() + ['data' => $dat->read()]);
+    }
+
+    /**
+     * Delete the referenced Dat by the `$id`.
+     *
+     * Currently it is a *soft* delete, so the database row nor the
+     * `.dat` file are really deleted.
+     *
+     * @param int $id
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function delete($id)
+    {
+        /**
+         * The Dat file.
+         *
+         * @var Dat|null $dat
+         */
+        $dat = Dat::find($id);
+
+        if(!$dat || $dat->deleted_at)
+        {
+            return response()->json(['error' => 'File not found'], 404);
+        }
+
+        $dat->deleted_at = Carbon::now();
+
+        if(!$dat->save())
+        {
+            return response()->json(['error' => 'File was not saved'], 500);
+        }
+
+        return response()->json(['deleted' => $dat->name]);
     }
 }