+++ bootstrap/app.php
@@ -49,6 +49,10 @@ $app->configure('jwt');
 |--------------------------------------------------------------------------
 */
 
+$app->middleware([
+    \App\Http\Middleware\ParsePUT::class
+]);
+
 $app->routeMiddleware([
     'auth' => \App\Http\Middleware\Authenticate::class,
     'file' => \App\Http\Middleware\HasFile::class