+++ routes/web.php
@@ -27,6 +27,7 @@ $router->group(['prefix' => 'api'], function() use ($router) {
     | Auth
     |----------------------------------------------------------
     */
+
     $router->post('register', 'AuthController@register');
     $router->post('login', 'AuthController@login');
     $router->post('logout', 'AuthController@logout');
@@ -37,8 +38,22 @@ $router->group(['prefix' => 'api'], function() use ($router) {
     | Dat
     |----------------------------------------------------------
     */
+
     $router->get('dat', 'DatController@list');
     $router->get('dat/{id}', 'DatController@read');
     $router->post('dat', 'DatController@store');
+    $router->put('dat/{id}', 'DatController@update');
+    $router->delete('dat/{id}', 'DatController@delete');
 
+
+    /*
+    |----------------------------------------------------------
+    | DoneDat
+    |----------------------------------------------------------
+    */
+
+    $router->get('dump', 'DoneDatController@list');
+    $router->get('dump/{id}', 'DoneDatController@read');
+    $router->post('dump', 'DoneDatController@dump');
+    $router->delete('dump/{id}', 'DoneDatController@delete');
 });